refactor: move new tests to unit_test folder
